+++ tests/test_save_dump.py
@@ -1,15 +1,24 @@
 import sys
 
+import pytest
+
 from debuglater import pydump
 
 
-def test_save_dump_without_dill(tmp_empty, monkeypatch):
+@pytest.mark.parametrize(
+    "path",
+    [
+        "name.dump",
+        "path/to/file.dump",
+    ],
+)
+def test_save_dump_without_dill(tmp_empty, monkeypatch, path):
     # simulate dill is not installed
-    monkeypatch.setattr(pydump, 'dill', None)
+    monkeypatch.setattr(pydump, "dill", None)
 
     try:
         1 / 0
     except Exception:
         tb = sys.exc_info()[2]
 
-    pydump.save_dump('name.dump', tb=tb)
+    pydump.save_dump(path, tb=tb)