+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 
 ## 1.4.4dev
+* `save_dump` creates all parent directories if they don't exist
 
 ## 1.4.3 (2022-08-22)
 * Adds `debug_dump` as top-level import